+++ docsig/_main.py
@@ -30,6 +30,9 @@ def main() -> int:
     members = _get_members(paths)
     for module in members:
         _populate(module.name, module, failures, missing)
+        for klass in module.classes:
+            name = f"{module.name}::{klass.name}"
+            _populate(name, klass, failures, missing)
 
     _print_failures(failures)
     _warn(missing)