+++ docsig/_module.py
@@ -15,10 +15,10 @@ class Parent:  # pylint: disable=too-few-public-methods
     :param node: Abstract syntax tree.
     """
 
-    def __init__(self, node: _ast.AST) -> None:
+    def __init__(self, node: _ast.AST, method: bool = False) -> None:
         self._node = node
         self._funcs = [
-            _Function(f)
+            _Function(f, method)
             for f in self._node.body  # type: ignore
             if isinstance(f, _ast.FunctionDef)
             and not str(f.name).startswith("_")
@@ -28,6 +28,21 @@ class Parent:  # pylint: disable=too-few-public-methods
     def funcs(self) -> _t.List[_Function]:
         """List of functions contained within the module."""
         return self._funcs
+
+
+class Class(Parent):  # pylint: disable=too-few-public-methods
+    """Represents a class with method parameters.
+
+    :param node: Abstract syntax tree.
+    """
+
+    def __init__(self, node: _ast.ClassDef) -> None:
+        super().__init__(node, method=True)
+
+    @property
+    def name(self) -> str:
+        """Name of module."""
+        return self._node.name  # type: ignore
 
 
 class Module(Parent):
@@ -40,8 +55,18 @@ class Module(Parent):
         node = _ast.parse(path.read_text(), filename=str(path))
         super().__init__(node)
         self._path = path
+        self._classes = [
+            Class(f)
+            for f in node.body  # type: ignore
+            if isinstance(f, _ast.ClassDef) and not str(f.name).startswith("_")
+        ]
 
     @property
     def name(self) -> str:
         """Name of module."""
         return str(self._path)
+
+    @property
+    def classes(self) -> _t.List[Class]:
+        """List of ``Class`` objects."""
+        return self._classes